+++ lib/mut/mutator/integer_literal.ex
@@ -29,11 +29,12 @@ defmodule Mut.Mutator.IntegerLiteral do
   def description, do: "Replace integer literal in function body."
 
   @impl true
-  def targets, do: [:body_literal]
+  def targets, do: [:body_literal, :pattern_literal]
 
   @impl true
   def applicable?(node, %Mut.Context{} = ctx) do
-    ctx.env_context == nil and ctx.engine in [:fallback, :schema] and integer_literal?(node)
+    ctx.env_context in [nil, :match] and ctx.engine in [:fallback, :schema] and
+      integer_literal?(node)
   end
 
   @impl true