+++ lib/mut/mutator/string_literal.ex
@@ -51,11 +51,11 @@ defmodule Mut.Mutator.StringLiteral do
   def description, do: "Replace non-empty string body literal."
 
   @impl true
-  def targets, do: [:env_walker]
+  def targets, do: [:env_walker, :pattern_literal]
 
   @impl true
   def applicable?(node, %Mut.Context{} = ctx) do
-    ctx.env_context == nil and ctx.engine in [:fallback, :schema] and
+    ctx.env_context in [nil, :match] and ctx.engine in [:fallback, :schema] and
       non_empty_string_literal?(node)
   end